HBASE-24635 Split TestMetaWithReplicas (#1980)
