+++ hbase-it/src/test/java/org/apache/hadoop/hbase/IntegrationTestMetaReplicas.java
@@ -18,9 +18,8 @@
 package org.apache.hadoop.hbase;
 
 import java.io.IOException;
-
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.hbase.client.TestMetaWithReplicas;
+import org.apache.hadoop.hbase.client.TestMetaWithReplicasShutdownHandling;
 import org.apache.hadoop.hbase.protobuf.ProtobufUtil;
 import org.apache.hadoop.hbase.regionserver.StorefileRefresherChore;
 import org.apache.hadoop.hbase.testclassification.IntegrationTests;
@@ -98,7 +97,7 @@ public class IntegrationTestMetaReplicas {
     // server holding the primary meta replica. Then it does a put/get into/from
     // the test table. The put/get operations would use the replicas to locate the
     // location of the test table's region
-    TestMetaWithReplicas.shutdownMetaAndDoValidations(util);
+    TestMetaWithReplicasShutdownHandling.shutdownMetaAndDoValidations(util);
   }
 
   public static void main(String[] args) throws Exception {
